rename logging/helpers to logging/fields
use our NoopLogger instead of logur.NoopLogger
