+++ server/events/command_runner.go
@@ -28,7 +28,7 @@ import (
 	"github.com/runatlantis/atlantis/server/events/models"
 	"github.com/runatlantis/atlantis/server/events/vcs"
 	"github.com/runatlantis/atlantis/server/logging"
-	logHelpers "github.com/runatlantis/atlantis/server/logging/helpers"
+	"github.com/runatlantis/atlantis/server/logging/fields"
 	"github.com/runatlantis/atlantis/server/recovery"
 	"github.com/uber-go/tally"
 	gitlab "github.com/xanzy/go-gitlab"
@@ -181,7 +181,8 @@ func (c *DefaultCommandRunner) RunAutoplanCommand(logger logging.SimpleLogging, 
 	}
 
 	if err := c.PreWorkflowHooksCommandRunner.RunPreHooks(context.TODO(), ctx); err != nil {
-		c.Logger.ErrorContext(context.TODO(), "Error running pre-workflow hooks", logHelpers.PullRequestWithErr(pull, err))
+		//TODO: thread context and use related logging methods.
+		c.Logger.ErrorContext(context.TODO(), "Error running pre-workflow hooks", fields.PullRequestWithErr(pull, err))
 		c.CommitStatusUpdater.UpdateCombined(context.TODO(), ctx.HeadRepo, ctx.Pull, models.FailedCommitStatus, command.Plan)
 		return
 	}
@@ -248,8 +249,8 @@ func (c *DefaultCommandRunner) RunCommentCommand(logger logging.SimpleLogging, b
 	}
 
 	if err := c.PreWorkflowHooksCommandRunner.RunPreHooks(context.TODO(), ctx); err != nil {
-		c.Logger.ErrorContext(context.TODO(), "Error running pre-workflow hooks", logHelpers.PullRequestWithErr(pull, err))
-
+		//TODO: thread context and use related logging methods.
+		c.Logger.ErrorContext(context.TODO(), "Error running pre-workflow hooks", fields.PullRequestWithErr(pull, err))
 		c.CommitStatusUpdater.UpdateCombined(context.TODO(), ctx.HeadRepo, ctx.Pull, models.FailedCommitStatus, cmd.Name)
 		return
 	}
